backend/requirements.txt
pydantic~=1.10.6
graphene~=3.2.2
fastapi~=0.95.0
uvicorn~=0.21.1
strawberry-graphql==0.168.1
Faker==18.3.2
requests~=2.28.2
starlette~=0.26.1
joblib~=1.2.0
scikit-learn~=1.2.2
pandas~=1.5.3
stix2~=3.0.1
taxii2-client~=2.3.0
beautifulsoup4~=4.12.0
attackcti~=0.3.8